switch upstream to "kingdonb/rvm-supported"
This upstream includes a global gemset with most of the gems we're
using, from every Ruby app, with every dependency precompiled.

Make sure you're not using the default gemset in any downstream, create
a gemset like 'testing' gemset that we've defined here.

Adding the testing gemset to login shells in .profile, for better use
with the Okteto CLI.