do not use a login shell
This fixes an issue where the terminal may report that we are using Ruby
2.6.6@testing gemset, twice on start. Invoking rvm RUBY@GEMSET does this
for us, while `bash --login` might – if we have also wrote this into
.profile (which we did):

RUN echo 'export PATH="$PATH:$HOME/app/bin"' >> ~/.profile \
 && echo "rvm ${RUBY}@${GEMSET}" >> ~/.profile

It's bad juju to use a login shell when you are not a login user.
So we don't do it in our example.